requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
passlib==1.7.4
bcrypt==4.0.1
python-jose ==3.3.0
python-multipart==0.0.5